Initial idea for supporting child frame's for windows.
Imgui allows the user to create child frame's, and render into them.
This isn't exposed currently, this is my first idea at how rust can
support child frame's.

Presently it's not the easiest to use, to have a window with a child
frame supported, internally the imgui library must have called
beginWindow() for the parent window, before beginChild() is ever called
for the child frame. Doing this without causing unneeded
allocations/complexity, and making the API ergonomic is something I hope
to work on next / get some feedback on.