Expose color variable stack to user.
In order to have colored and wrapped text, I needed to use this
functionality in my game. This PR implements a version similar to
with_style_var().

From my testing it is safe to use a Vec4 for all the variants of the
enumeration, so I didn't bother putting an enumeration in imgui-rs, I
just reused the enumeration defined in imgui-sys directly.

Let me know if you have any feedback on this.